Audio system improved, settings ui fix, LCD fix
- Added new strip mode
- Improved average volume trigger level
- Fixed Settings UI strip selection mode not flashing white
- Fixed LCD

+++ ZZ/code/common/inc/FreeRTOS_def_decl.h
@@ -10,7 +10,7 @@
 /************************************************************************/
 
 #ifndef DEBUG
-	#define delayFREERTOS(x)	vTaskDelay(x/portTICK_PERIOD_MS);
+	#define delayFREERTOS(x)	vTaskDelay((double)x/portTICK_PERIOD_MS);
 #else
 	#include <math.h>
 	#define delayFREERTOS(x)	vTaskDelay(			2		)
